+++ src/app/components/collab/room-panel.component.ts
@@ -53,6 +53,11 @@ export class RoomPanelComponent implements OnInit {
     });
 
     this.userService.currentUser$.subscribe((user) => {
+
+      if (!this.canCollaborate) {
+        return;
+      }
+
       const roomInfo = this.roomInfoArray.find(info => info.docId === this.activeDocId);
       // If user changed or logout, leave room first
       if (roomInfo && roomInfo.joinedRoom) {
@@ -65,6 +70,12 @@ export class RoomPanelComponent implements OnInit {
     });
 
     this.collabService.roomParticipantsChange$.subscribe((roomParticipants: RoomParticipants) => {
+
+      if (!this.canCollaborate) {
+        return;
+      }
+
+
       // if the change is for active doc, then handle it. Otherwise, ignore it.
       const roomInfo = this.roomInfoArray.find(info => info.docId === this.activeDocId);
       if (roomInfo) {
@@ -144,6 +155,12 @@ export class RoomPanelComponent implements OnInit {
   }
 
   updateRoomByCurrentFile() {
+
+    if (!this.canCollaborate) {
+      return;
+    }
+
+
     let fileInfo = RXCore.getCurrentFileInfo();
 
     if(fileInfo != undefined){